Replace obsolete AC_TRY_FOO macros with AC_FOO_IFELSE
Autoconf made several macros obsolete including the AC_TRY_COMPILE and
AC_TRY_RUN in 2000:
http://git.savannah.gnu.org/cgit/autoconf.git/tree/ChangeLog.2

It should be replaced with the current AC_COMPILE_IFELSE and the
AC_RUN_IFELS instead:
- AC_TRY_COMPILE with AC_COMPILE_IFELSE and AC_LANG_PROGRAM
- AC_TRY_RUN with AC_RUN_IFELSE and AC_LANG_SOURCE

It is fairly safe to upgrade and take the recommendation advice of
autoconf upgrade manual since the upgrade should be compatible at least
with PHP versions 5.4 and up, on some systems even with PHP 5.3.

For example, PHP 5.4 to 7.1 require Autoconf 2.59+ version, PHP 7.2 and
above require 2.64+ version, and the PHP 7.2 phpize script requires 2.59+
version which are all greater than above mentioned 2.50 version.

This patch was created with the help of autoupdate script:
autoupdate timelib.m4

Reference docs:
- https://www.gnu.org/software/autoconf/manual/autoconf-2.69/html_node/Obsolete-Macros.html
- https://www.gnu.org/software/autoconf/manual/autoconf-2.69/html_node/Autoconf-2_002e13.html