Support timezones where transitions don't include all POSIX abbreviations
The America/Indiana/Petersburg and America/North_Dakota/Beulah zones define a
POSIX string with a dst abbreviation that was not used in any other transition,
and hence, these abbreviations were not stored in the 'timezone_abbr' structure
in the TZ file. With this patch, these new abbreviations are now added to that
string, and new 'type' structures are created as well.